containers: add symlink resolution support to path resolver
- Add ResolveLink() method for single symlink resolution with loop detection
- Add ResolveAllLinks() method for comprehensive path component resolution
- Add GetProcMounts() method to find accessible mount info in namespaces
- Implement LRU cache (1024 entries) for symlink resolution performance
- Add security validation to prevent symlink traversal outside mount boundaries
- Change mount namespace type from int to uint32 for consistency
- Add comprehensive input validation for all new methods
- Update related types and interfaces for type consistency
- Add extensive test coverage for new functionality and edge cases

This enables tracee to properly resolve symlinks within container mount
namespaces while maintaining security and performance.

+++ pkg/ebpf/processor_funcs.go
@@ -197,7 +197,7 @@ func (t *Tracee) processSchedProcessExec(event *trace.Event) error {
 					return errfmt.Errorf("error parsing sched_process_exec args: %v", err)
 				}
 
-				fileKey := filehash.NewKey(filePath, event.MountNS,
+				fileKey := filehash.NewKey(filePath, uint32(event.MountNS),
 					filehash.WithDevice(dev),
 					filehash.WithInode(ino, castedSourceFileCtime),
 					filehash.WithDigest(event.Container.ImageDigest),
@@ -409,7 +409,7 @@ func (t *Tracee) processSharedObjectLoaded(event *trace.Event) error {
 		containerId = "host"
 	}
 	if t.config.Output.CalcHashes != config.CalcHashesNone {
-		fileKey := filehash.NewKey(filePath, event.MountNS,
+		fileKey := filehash.NewKey(filePath, uint32(event.MountNS),
 			filehash.WithDevice(dev),
 			filehash.WithInode(ino, int64(fileCtime)),
 			filehash.WithDigest(event.Container.ImageDigest),
